Finish off averaging data work
Adds code which will take a list of series objects and will average the
contents if it contains 2 min/max columns for a piece of data and will
create a series object using this. It will then take all of these and
join them together into a dataframe which is then returned.