Update tau based on average of cluster losses
Update tau based on a moving average of the output cluster information losses.
If no clusters have been output yet, then use the average information loss of a
sample of the clusters in Γ. If no information has been entered into CASTLE
yet, then a value of positive infinity is used.