Add optional seed and sample size arguments
Add optional arguments to the CLI for specifying a seed and sample size to the
simulation. This allows for easier debugging, as not all runs produce a crash.
The simulation will display the seed when it picks it, and running with --seed
SEED will allow the simulation to be re-run exactly.

For example, running with --seed 100 currently will cause a crash, but --seed
101 will not. Reproduceability is important for figuring out why certain things
work and others do not.