Fix tuple suppression to work correctly
When a tuple is suppressed, it should be removed from the global tuple queue,
which is what was happening before correctly. It should also however remove the
tuple from the cluster it is being contained in. Otherwise, the cluster may be
output later, and it will try and output this tuple for a second time. This
will cause an error when attempting to remove the tuple from the queue, as it
has already been ejected.