Ensure Items cannot be in multiple clusters
Ensure Items cannot be contained within multiple clusters. Cluster.insert
implements a basic detector, checking that the inserted item does not have a
parent. If it does, it removes it from that cluster and then updates the parent
node. Add a test to ensure this is the case as well.