Fix split_l and generate_buckets
Fix the split_l function in CASTLE, along with generate_buckets, and return the
args.sample_size variable to its rightful place.

np.random.randint is non-inclusive unlike random.randint, so we need to remove
the -1.

When calculating the number of tuples to use, we should truncate it as it can
sometimes be a float. This is inline with what Alistair suggested.

Item objects cannot access the underlying data through a `.`, so `[attr]` must
be used instead, and Clusters do not allow for iteration, so this should be
C.contents.

generate_buckets needed a docstring, so I added one. It also was only making a
single bucket as there was only ever a single sensitive attribute, where it
should be making one for each value that the sensitive attribute has in the
data.

Return args.sample_size to where it should be instead of always using a
subsample of 20.