Update total_tuples to length of a set
Update the `total_tuples` variable in `delay_constraint` to be the length of
the set of all `pid` values, instead of the length of each cluster
individually. This is because otherwise we can reach `merge_clusters` without
enough unique `pid` values to satisfy the `k-anonymity` constraint, at which
point we will get a `min() arg is an empty sequence` error.